Add search functionality with lunr.js
The lunr.js library has been added as a dependency. It is used in the server.js file to create an indexed search system, replacing the previous TODO. This system efficiently searches game titles, descriptions, and developers. A console.log to check game data has also been added in the search.mustache file.
We will be adding the search bar shortly.