Update game link navigation and search fields
Modified the onclick attribute in search.mustache to correctly redirect users to a game's page. The path to the game page is now absolute, preventing the path from repeatedly appending on each click. Also, removed the 'text' and 'developer' fields from the lunr search index in server.js as they are not required for searching games.